Add Wasm publish base test class
